requirements.txt
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
seaborn==0.13.2
kafka-python==2.0.2
six==1.16.0
pyspark==3.5.3
setuptools==75.3.0
xlrd==2.0.1
openpyxl==3.1.5
